Add files via upload
The testClientForWCFsoap is a console app that is a poor man's substitute for a web client. It calls the services, which do some messaging stuff synch. Ideally it should be async and use a call back to the client once processing is done.
Note: I was tinkering with NUnit for project testing. I did not include those packages so you will have to download them yourself.